+++ include/arion/components/arion_afl.hpp
@@ -45,7 +45,7 @@ class ARION_EXPORT ArionAfl
     ARION_EXPORT ArionAfl(std::weak_ptr<Arion> arion) : arion(arion) {};
     void ARION_EXPORT fuzz(ARION_AFL_INPUT_CALLBACK input_callback, ARION_AFL_CRASH_CALLBACK crash_callback,
                            std::vector<arion::ADDR> exits, bool keep_mem = true,
-                           std::vector<int> signals = {SIGSEGV, SIGABRT}, bool always_validate = true,
+                           std::vector<int> signals = {SIGSEGV, SIGABRT}, bool always_validate = false,
                            uint32_t persistent_iters = 1000, void *user_data = nullptr);
 };